+++ packages/@cosmos-building-blocks/service/src/ecs/ecs-service.ts
@@ -24,15 +24,22 @@ import {
   ApplicationListenerRuleProps,
   ApplicationTargetGroupProps,
   ListenerAction,
+  ILoadBalancerV2,
+  ListenerCondition,
 } from '@aws-cdk/aws-elasticloadbalancingv2';
 import { IVpc } from '@aws-cdk/aws-ec2';
 import { LogGroup, LogGroupProps, RetentionDays } from '@aws-cdk/aws-logs';
 import { EnableScalingProps } from '@aws-cdk/aws-applicationautoscaling';
-import { getRoutingPriorityFromListenerProps } from '../utils';
+import { ARecord, IHostedZone, RecordTarget } from '@aws-cdk/aws-route53';
+import { LoadBalancerTarget } from '@aws-cdk/aws-route53-targets';
+import { Certificate, DnsValidatedCertificate } from '@aws-cdk/aws-certificatemanager';
+import { getRoutingPriorityFromListenerProps } from '.';
 
 export interface EcsServiceProps {
   vpc: IVpc;
+  zone?: IHostedZone;
   cluster: ICluster;
+  alb?: ILoadBalancerV2;
   httpListener?: IApplicationListener;
   httpsListener?: IApplicationListener;
   logProps?: Partial<LogGroupProps>;
@@ -40,9 +47,13 @@ export interface EcsServiceProps {
   containerProps: ContainerDefinitionOptions & { port?: PortMapping };
   serviceProps?: Partial<Ec2ServiceProps>;
   targetGroupProps?: Partial<ApplicationTargetGroupProps>;
-  routingProps?: Partial<ApplicationListenerRuleProps>;
+  routingProps?: Partial<ApplicationListenerRuleProps> & {
+    httpsRedirect?: boolean;
+    subdomains?: string[];
+    certificate?: boolean;
+    dns?: boolean;
+  };
   scalingProps?: EnableScalingProps;
-  httpsRedirect?: boolean;
 }
 
 export class EcsService extends Construct {
@@ -51,15 +62,19 @@ export class EcsService extends Construct {
   readonly container: ContainerDefinition;
   readonly service: Ec2Service;
   readonly targetGroup?: ApplicationTargetGroup;
-  readonly listenerRules?: ApplicationListenerRule[];
+  readonly listenerRules: ApplicationListenerRule[];
   readonly scaling?: ScalableTaskCount;
+  readonly certificate?: Certificate;
+  readonly subdomains: ARecord[];
 
   constructor(scope: Construct, id: string, props: EcsServiceProps) {
     super(scope, id);
 
     const {
       vpc,
+      zone,
       cluster,
+      alb,
       httpListener,
       httpsListener,
       logProps,
@@ -69,8 +84,10 @@ export class EcsService extends Construct {
       targetGroupProps,
       routingProps,
       scalingProps,
-      httpsRedirect,
     } = props;
+
+    this.listenerRules = [];
+    this.subdomains = [];
 
     this.logGroup = new LogGroup(this, 'Logs', {
       retention: RetentionDays.ONE_MONTH,
@@ -102,6 +119,15 @@ export class EcsService extends Construct {
     });
 
     if (routingProps) {
+      const { httpsRedirect, subdomains, certificate, dns } = routingProps;
+
+      const conditions = [...(routingProps.conditions || [])];
+      const _routingProps = {
+        priority: 0,
+        ...routingProps,
+        conditions: conditions,
+      };
+
       if (!httpListener && !httpsListener) throw new Error('To enable routing, an Http Listener is required');
 
       this.targetGroup = new ApplicationTargetGroup(this, 'ServiceTargetGroup', {
@@ -119,13 +145,45 @@ export class EcsService extends Construct {
         ],
       });
 
-      this.listenerRules = [];
+      if (subdomains?.length) {
+        if (!zone) throw new Error('Please provide zone prop.');
+
+        conditions.push(ListenerCondition.hostHeaders(subdomains.map((x) => `${x}.${zone.zoneName}`)));
+
+        if (dns) {
+          if (!alb) throw new Error('Please provide alb prop.');
+
+          subdomains.forEach((subdomain, i) => {
+            const record = new ARecord(this, `Subdomain${i}`, {
+              zone: zone,
+              recordName: subdomain,
+              target: RecordTarget.fromAlias(new LoadBalancerTarget(alb)),
+            });
+            this.subdomains.push(record);
+          });
+        }
+
+        if (certificate) {
+          if (!alb) throw new Error('Please provide alb prop.');
+          if (!httpsListener) throw new Error('Please provide httpsListener prop.');
+
+          this.certificate = new DnsValidatedCertificate(this, 'Certificate', {
+            hostedZone: zone,
+            domainName: zone.zoneName,
+            subjectAlternativeNames: subdomains.map((x) => `${x}.${zone.zoneName}`),
+          });
+
+          httpsListener.addCertificateArns(this.certificate.node.id, [this.certificate.certificateArn]);
+        }
+      }
+
+      // Render the priority
+      _routingProps.priority = getRoutingPriorityFromListenerProps(this, _routingProps);
 
       if (httpsListener) {
         this.listenerRules.push(
           new ApplicationListenerRule(this, `HttpsServiceRule`, {
-            priority: getRoutingPriorityFromListenerProps(routingProps),
-            ...routingProps,
+            ..._routingProps,
             listener: httpsListener,
             action: ListenerAction.forward([this.targetGroup]),
           })
@@ -136,8 +194,7 @@ export class EcsService extends Construct {
         if (!httpListener) throw new Error('To enable https redirect you must provide an httpListener');
 
         new ApplicationListenerRule(this, 'HttpsRedirect', {
-          priority: getRoutingPriorityFromListenerProps(routingProps),
-          ...routingProps,
+          ..._routingProps,
           listener: httpListener,
           action: ListenerAction.redirect({
             permanent: true,
@@ -148,8 +205,7 @@ export class EcsService extends Construct {
       } else if (httpListener) {
         this.listenerRules.push(
           new ApplicationListenerRule(this, `HttpServiceRule`, {
-            priority: getRoutingPriorityFromListenerProps(routingProps),
-            ...routingProps,
+            ..._routingProps,
             listener: httpListener,
             action: ListenerAction.forward([this.targetGroup]),
           })